% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cell_hierarchy}
\alias{cell_hierarchy}
\title{Create cell hierarchy}
\usage{
cell_hierarchy(counts, celltypes, counts.bulk = NULL)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A data frame containing the cell hiearchy. Each column describes the node lineage for a given cell-type.
}
\description{
Create cell hierarchy
}
