% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_organ_df}
\alias{cell_organ_df}
\title{Possible cell-types per organ}
\format{
A cell-type x organ data frame
}
\usage{
cell_organ_df
}
\description{
A data frame describing the organs in which each cell-type may occur. A value of '1' indicates the cell-type was directly detected in the given organ in Tabula Sapiens. A value of '2' indicates that the cell-type may be detected in other organs but was not detected in the Tabula Sapiens. This generally applies to most immune, stromal, and endothelial cell-types.
}
\keyword{datasets}
