% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_cancer_models}
\alias{census_cancer_models}
\title{Census cancer models}
\format{
A list of lists for each model. The model list format is the same as ts.model
}
\usage{
census_cancer_models
}
\description{
Collection of models to detect cancer cells in the following organs: breast, colon, kidney, liver, lung, pancreas
}
\keyword{datasets}
