% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CENSUS_MAX_VARIABLES}
\alias{CENSUS_MAX_VARIABLES}
\title{Maximum variables per Census data query}
\format{
A \verb{scalar<integer>}.
}
\usage{
CENSUS_MAX_VARIABLES
}
\description{
The Census Data API caps the \verb{get=} list at 50 variables per call.
}
\keyword{datasets}
