% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census_predict}
\alias{census_predict}
\title{Use a Census model to annotate new datasets}
\usage{
census_predict(obj, model, contour_data = F, verbose = T, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Use a Census model to annotate new datasets
}
