% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_ts_model}
\alias{census_ts_model}
\title{Census Tabula Sapiens model}
\format{
A list containing standard model outputs: models (xgboost tree model for each node), markers (marker genes used in each model), auc (model performance on training data), hierarchy_mat (cell-type hierarchy used by model)
}
\usage{
census_ts_model
}
\description{
Hierarchical gradient-boosted tree model trained on the Tabula Sapiens
}
\keyword{datasets}
