% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensus.R
\name{getCensus}
\alias{getCensus}
\title{Scrape US Census Data}
\usage{
getCensus(cities, radius)
}
\arguments{
\item{cities}{list of cities to be passed}

\item{radius}{numeric number used to create a boundary around the city coordinates}
}
\value{
Data frame containing demographic data
}
\description{
Takes in a city/state and outputs demographic data from the Missouri Census Data Center
}
\examples{
radius <- 2
cities <- c("Chicago", "New York")
df <- getCensus(cities, radius)
head(df)
}
