% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatLon.R
\name{getLatLon}
\alias{getLatLon}
\title{Get latidude and longitude}
\usage{
getLatLon(cities)
}
\value{
Demographic data
}
\description{
Takes in a city/state and outputs the name of the city, latitude and longitude
}
\examples{
cities <- c("Chicago", "New York")
getLatLon(cities)
}
