% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_cell_hierarchy}
\alias{plot_cell_hierarchy}
\title{Plot the cell hierarchy}
\usage{
plot_cell_hierarchy(
  hierarchy_mat,
  circular = F,
  color = c("#000004FF", "#420A68FF", "#DD513AFF", "#FCA50AFF"),
  edge.thickness = 0.5,
  text.size = 2,
  label.size = 2,
  label.line.padding = 0.25,
  axis.limits = NULL
)
}
\arguments{
\item{hierarchy_mat}{Cell-type hierarchy dataframe produced by the cell_hierarchy function}

\item{circular}{Plots a circular dendorgram if True}

\item{color}{Colors to use for cell lables. Can be a single color or multiple colors. If multiple colors are given they will be matched the major hierarchy branches.}

\item{edge.thickness}{Thickness of dendrogram lines}

\item{text.size}{Cell-type label text size}

\item{label.size}{Node label text size}

\item{label.line.padding}{Controls size of the node label outline}

\item{axis.limits}{Controls horizontal size of dendrogram. Takes arguements in the form "c(x,y)" where x and y can be positive or negative numbers}
}
\value{
A ggraph object plotting the cell-type hiearchy
}
\description{
Plot the cell hierarchy
}
