% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_node}
\alias{predict_node}
\title{Predict cell identities for a specific node in the hierarchical model}
\usage{
predict_node(obj, model, node, get_prob = T, allowed_nodes = NULL)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Predict cell identities for a specific node in the hierarchical model
}
