% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_node}
\alias{train_node}
\title{Train a classifier on a specific node in the cell hierarchy}
\usage{
train_node(
  obj,
  node,
  hierarchy_mat,
  celltypes,
  metadata = NULL,
  markers = NULL,
  cell_node_match = NULL,
  sparsity = c(0.9),
  noise = c(-0.5, 0.5),
  noise_frac = 0.5,
  ptile = 100,
  topn = NULL,
  nrounds = 20,
  eval_metric = "auc",
  colsample = 0.2,
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Train a classifier on a specific node in the cell hierarchy
}
