% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cell_hierarchy}
\alias{cell_hierarchy}
\title{Create cell hierarchy}
\usage{
cell_hierarchy(counts, celltypes, counts.bulk = NULL)
}
\arguments{
\item{counts}{Gene by cell counts matrix}

\item{celltypes}{Character vector of cell-type for each barcode in counts}

\item{counts.bulk}{Optional pseudo-bulk cell-type by gene counts matrix. May be useful if counts is very large and hard to transpose}
}
\value{
A data frame containing the cell hiearchy. Each column describes the node lineage for a given cell-type.
}
\description{
Create cell hierarchy
}
