% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census_main}
\alias{census_main}
\title{Automated cell-type annotation using the main Census model trained on the Tabula Sapiens and Cancer Cell Line Encyclopedia}
\usage{
census_main(
  obj,
  organ = NULL,
  test = "all",
  predict_cancer = F,
  contour_data = F,
  verbose = T,
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Automated cell-type annotation using the main Census model trained on the Tabula Sapiens and Cancer Cell Line Encyclopedia
}
