% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census_predict}
\alias{census_predict}
\title{Use a Census model to annotate new datasets}
\usage{
census_predict(obj, model, contour_data = F, verbose = T, ...)
}
\arguments{
\item{obj}{Seurat object containing training data}

\item{model}{Hierarchical models list that is the output of census_train. Contains the models, markers, and cell-type hierarchy}

\item{contour_data}{If True, contour data are outputed}

\item{verbose}{if True, messages are printed to the console during prediction.}

\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Use a Census model to annotate new datasets
}
