% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census_predict}
\alias{census_predict}
\title{Automated cell-type annotation using the main Census model}
\usage{
census_predict(
  obj,
  organ = NULL,
  test = "all",
  predict_cancer = F,
  contour_data = F,
  verbose = T,
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Automated cell-type annotation using the main Census model
}
