% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census_train}
\alias{census_train}
\title{Train a hierarchical Census model}
\usage{
census_train(
  obj,
  hierarchy_mat,
  celltypes,
  metadata = NULL,
  markers.list = NULL,
  sparsity = c(0.9),
  noise = c(-0.5, 0.5),
  noise_frac = 0.5,
  ptile = 100,
  nrounds = 20,
  eval_metric = "auc",
  colsample = 0.2,
  print_node = T,
  ...
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Train a hierarchical Census model
}
