% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contour_adjust}
\alias{contour_adjust}
\title{Label stabilizing algorithm using cluster and prediction contour adjustment}
\usage{
contour_adjust(
  pred_df,
  contour_level = 0.999,
  min_prob_diff = 0,
  contour_n = 200
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Label stabilizing algorithm using cluster and prediction contour adjustment
}
