% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_cell_hierarchy}
\alias{plot_cell_hierarchy}
\title{Plot the cell hierarchy}
\usage{
plot_cell_hierarchy(
  hierarchy.mat,
  circular = F,
  n.color.per.group = NULL,
  color = c("firebrick1", "firebrick4", "dodgerblue1", "dodgerblue4", "seagreen3",
    "darkgreen", "slateblue1", "slateblue4"),
  edge.thickness = 0.5,
  text.size = 2,
  label.size = 2,
  label.line.padding = 0.25,
  axis.limits = NULL
)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
A ggraph object plotting the cell hiearchy
}
\description{
Plot the cell hierarchy
}
